Fix googletest cmake build
Possibly due to changes in more recent versions of Googletest, or
possibly due to changes in more recent versions of CMake, the previous
way that CMake was invoked to build the googletest library failed with
errors about `GOOGLETEST_VERSION` being undefined. This was happening
because `cmake` was being run from `tests/googletest/googletest/make`
instead of `tests/googletest/googletest`. Changing the directory where
CMake is invoked and adding the `-B` flag fixes that problem.