Rename and improve python_format.yml
This renames `python_format.yml` to `ci_format_checks.yaml`, to make
the purpose more clear from the file name and to allow for the
possibility that C++ format checks will be added later.

In addition, this makes the workflow check whether the changes in a PR
affected Python files. If not, execution is skipped, for greater
efficiency.

This also adds a `workflow_dispatch` event trigger so that this can be
invoked manually (e.g., for testing).